sibyl_matrix: Add support for joining rooms on invite (fixes #38)
By default we still ignore room invites, but this adds the option to
automatically accept invites from either all users or users on the
same homeserver. This is necessary to allow sibyl to respond to PMs in
Matrix, as Matrix doesn't have the concept of 1-1 chats (all chat
happens within a room that you're first invited to).