Ignore messages sent before joining a room (fixes #39)
Matrix returns historical messages in the next /sync after joining a
room. To avoid this, we can keep track of when we joined each room and
make sure the message timestamp comes after that.

This could result in ignoring messages if the originating homeserver's
clock isn't synced (or if the client thinks it's in the future).