Delete saved creds on auth failure and rename creds var
In the event an access token is revoked, we might still have user/pass
credentials that are valid. If we receive a 401 or 403, we can be
reasonably sure that the credentials in the bot's state are invalid,
so let's just clear them out.

Sibyl will still disable the Matrix protocol if this happens,
requiring a restart; this just ensures that you won't have to dig
around in/delete the pickled state to get back up and running.