Added game-playing functionality using bot commands.
Added the [game], [register], [unregister], [startgame], [quitgame], and [move] commands. Added a generic Game object that gets instantiated to None on bot startup. Added a classes file for containing Game classes, which currently consists of a single game, Battleship. The Game object is designed to be generic enough to support any type of multiplayer game that can be played in a chatroom, with more game-specific settings able to be specified in the classes file. Games are played using either a public chatroom, private messages, or both.